Added autoIncriminating id for Player.java

+++ src/main/java/com/project/chicagoscoretracker/controller/PlayerController.java
@@ -33,6 +33,6 @@ public class PlayerController {
     public Player getPlayerByName(String name){return service.getPlayerByName(name);}
 
     @PostMapping("/player/delete")
-    public void deletePlayer(@RequestBody int id){service.deletePlayer(id);}
+    public void deletePlayer(@RequestBody Long id){service.deletePlayer(id);}
 
 }
